backend/requirements.txt
# requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
bcrypt
pydantic